File moved to Debugging -> Errors .h/cpp
