+++ src/collision/Maps/MapTree.cpp
@@ -22,7 +22,7 @@
 #include "VMapManager2.h"
 #include "VMapDefinitions.h"
 #include "Logging/Logger.hpp"
-#include "Errors.h"
+#include "Debugging/Errors.h"
 
 #include <string>
 #include <sstream>
