+++ src/shared/Threading/LegacyThreading.h
@@ -22,7 +22,7 @@
 #define _THREADING_H
 
 // We need assertions.
-#include <Errors.h>
+#include "Debugging/Errors.h"
 
 // Platform Specific Lock Implementation
 #include "Mutex.h"
