+++ src/world/Movement/Spline/Spline.h
@@ -6,7 +6,7 @@ This file is released under the MIT license. See README-MIT for more information
 #pragma once
 
 #include "MovementTypedefs.h"
-#include "Errors.h"
+#include "Debugging/Errors.h"
 #include <G3D/Vector3.h>
 #include <limits>
 #include <vector>
